+++ packages/devtools_app/lib/src/inspector/flutter/layout_explorer/free_space.dart
