+++ packages/devtools_app/lib/src/inspector/flutter/layout_explorer/overflow_indicator_painter.dart
