+++ packages/devtools_app/lib/src/inspector/flutter/layout_explorer/utils.dart
