+++ packages/devtools_app/test/flutter/layout_explorer/goldens/arrow_bidirectional_horizontal.png
