+++ packages/devtools_app/test/flutter/layout_explorer/goldens/arrow_unidirectional_up.png
