+++ packages/devtools_app/test/flutter/layout_explorer/goldens/story_of_column_layout.png
