+++ packages/devtools_app/test/flutter/layout_explorer/goldens/story_of_row_layout.png
