+++ packages/devtools_app/test/flutter/layout_explorer/inspector_data_models_test.dart
@@ -6,7 +6,7 @@ import 'dart:convert';
 
 import 'package:devtools_app/src/inspector/diagnostics_node.dart';
 import 'package:devtools_app/src/inspector/flutter/inspector_data_models.dart';
-import 'package:devtools_app/src/inspector/flutter/story_of_your_layout/utils.dart';
+import 'package:devtools_app/src/inspector/flutter/layout_explorer/utils.dart';
 import 'package:flutter/widgets.dart';
 import 'package:test/test.dart';
