+++ packages/devtools_app/test/flutter/layout_explorer/utils_test.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:devtools_app/src/inspector/flutter/story_of_your_layout/utils.dart';
+import 'package:devtools_app/src/inspector/flutter/layout_explorer/utils.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
